% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{delete_signal}
\alias{delete_signal}
\title{Delete Signal}
\usage{
delete_signal(token, signal_id)
}
\arguments{
\item{token}{User access token}

\item{signal_id}{Signal ID}
}
\value{
HTTP response
}
\description{
deletes a signal
}
