% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{get_signal}
\alias{get_signal}
\title{Get Signal}
\usage{
get_signal(access_token, signal_id, infer_types = TRUE)
}
\arguments{
\item{access_token}{User's access token}

\item{signal_id}{Signal ID}

\item{infer.types}{Whether to infer columnn data types (defaults to TRUE)}
}
\value{
A data.frame containing the data for a signal
}
\description{
returns a signal's data in data.frame format
}
