% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{get_signal_details}
\alias{get_signal_details}
\title{List Signal Details}
\usage{
get_signal_details(access_token, signal_id)
}
\arguments{
\item{access_token}{User's access token}

\item{signal_id}{Signal ID}
}
\value{
A data.frame containing the details for a given signal
}
\description{
Function Description
}
