% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{signal_to_csv}
\alias{signal_to_csv}
\title{Save Signal to CSV}
\usage{
signal_to_csv(access_token, signal_id, file_name)
}
\arguments{
\item{access_token}{User's access token}

\item{signal_id}{Signal ID}

\item{file_name}{File name for the CSV}
}
\description{
Save Signal to CSV
}
