% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{auto_discover}
\alias{auto_discover}
\title{Auto Discover}
\usage{
auto_discover(
  token,
  geo_grain,
  filename = NULL,
  df = NULL,
  callback_url = NULL
)
}
\arguments{
\item{token}{User access token}

\item{geo_grain}{Geo-grain of upload, "State" or "Country"}

\item{filename}{Filename of .CSV or .XLS with "Date", "Value", "State"
(if geo_grain=State) columns, not to be used with \code{df}}

\item{df}{DataFrame with "Date", "Value", "State" (if geo_grain=State),
not to be used with \code{filename}}

\item{callback_url}{Callback URL for notifications}
}
\value{
HTTP response
}
\description{
create a signal with the Auto Discover feature
}
