% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{auto_discover}
\alias{auto_discover}
\title{Auto Discover}
\usage{
auto_discover(
  token,
  geo_grain,
  date_grain,
  filename = NULL,
  df = NULL,
  create_custom_features = 1,
  callback_url = NULL
)
}
\arguments{
\item{token}{User access token}

\item{geo_grain}{Geo grain of upload, "State" or "Country"}

\item{date_grain}{Date grain of upload, "Day" or "Month"}

\item{filename}{Filename of .CSV or .XLS with "Date", "Value", "State"
(if geo_grain=State) columns. Not to be used with \code{df}}

\item{df}{DataFrame with "Date", "Value", "State" (if geo_grain=State).
Not to be used with \code{filename}}

\item{create_custom_features}{Flag for custom feature creation. 0 or 1, defaults to 1}

\item{callback_url}{Callback URL for notifications}
}
\value{
HTTP response
}
\description{
create a signal with the Auto Discover feature
}
