% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{get_signal}
\alias{get_signal}
\title{Get Signal}
\usage{
get_signal(access_token, signal_id)
}
\arguments{
\item{access_token}{User's access token}

\item{signal_id}{Signal ID}
}
\value{
A data.frame containing the data for a signal
}
\description{
Function Description
}
