% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{get_signal}
\alias{get_signal}
\title{Get Signal}
\usage{
get_signal(token, signal_id, infer_types = TRUE)
}
\arguments{
\item{token}{User access token}

\item{signal_id}{Signal ID}

\item{infer_types}{Whether to infer column data types (defaults to TRUE)}
}
\value{
A data.frame containing the data for a signal
}
\description{
returns a signal's data in data.frame format
}
