% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{list_signals}
\alias{list_signals}
\title{List Signals}
\usage{
list_signals(access_token)
}
\arguments{
\item{access_token}{User's access token}
}
\value{
A data.frame containing the list of signals
}
\description{
Function Description
}
