% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{list_signals}
\alias{list_signals}
\title{List Signals}
\usage{
list_signals(token)
}
\arguments{
\item{token}{User access token}
}
\value{
A data.frame containing the list of signals
}
\description{
lists all the signals associated with the User access token
}
