% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readysignal.R
\name{signal_to_csv}
\alias{signal_to_csv}
\title{Save Signal to CSV}
\usage{
signal_to_csv(token, signal_id, file_name)
}
\arguments{
\item{token}{User access token}

\item{signal_id}{Signal ID}

\item{file_name}{File name for the CSV}
}
\value{
None, function writes data to file system
}
\description{
saves signal data to CSV file
}
